Refactor duplicate PasswordWrapper
